requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
pandas>=2.0.0
xgboost>=1.7.0
scipy>=1.10.0
scikit-learn>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
openpyxl>=3.0.0
httpx>=0.24.0
pyyaml>=6.0

# RAG相关依赖
faiss-cpu>=1.7.0  # 或 faiss-gpu（如果有GPU）
numpy>=1.24.0
dashscope>=1.14.0  # DashScope SDK（用于embedding和rerank API）

# 认证相关依赖
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
aiosqlite>=0.16.0
email-validator>=1.3.0
bcrypt<4.1

# 地图可视化相关依赖
shapely>=2.0.0
geopandas>=0.13.0
fiona>=1.9.0

# 可选依赖（本地embedding模式需要）
# FlagEmbedding>=1.2.0  # 取消注释以启用本地BGE-M3模型
